+++ src/indieweb/__init__.py
@@ -2,4 +2,4 @@
 includes indieauth and a micropub endpoint
 """
 
-__version__ = "0.3.4"
+__version__ = "0.3.5"